Pointers
Changed Units and Cities to be pointers and have memory freed, also added crude test to see if a city can be reached from another

+++ src/Unit.cpp
@@ -1,5 +1,5 @@
 #include "Unit.h"
 
 void Unit::print() const{
-    printf ("Unit #%zu: Belong to %zu with combat %d, can move %d, rebase:%d, and landing:%d\n", id, nationality, combat_value, movement, rebase, landing);
+    printf ("Unit #%zu: Belong to %d with combat %d, can move %d, rebase:%d, and landing:%d\n", id, (int)nationality, (int)combat_value, (int)movement, (int)rebase, (int)landing);
 }
